implement admin assign project manger for  project with validation of user(project managment role) id and it is role, and the project exsistance

+++ src/main/java/com/orienteed/orienteed/management/system/Repository/userRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface userRepo extends JpaRepository<User, Long> {
+public interface userRepo extends JpaRepository<User, Integer> {
 
 
     User findByuserName(String username);
